FIX: narrowing conversions

+++ src/font_decomposition/EdgeSegment.h
@@ -17,10 +17,10 @@ public:
     
     virtual ~EdgeSegment() = default;
     
-    [[nodiscard]] virtual float Distance(const Vector2& p, double& t) const = 0;
-    [[nodiscard]] virtual float PseudoDistance(float distance, const Vector2&p, double& t) const = 0;
+    [[nodiscard]] virtual double Distance(const Vector2& p, double& t) const = 0;
+    [[nodiscard]] virtual double PseudoDistance(float distance, const Vector2&p, double& t) const = 0;
     
-    [[nodiscard]] virtual float SignedDistance(const Vector2& p, double& t) const;
+    [[nodiscard]] virtual double SignedDistance(const Vector2& p, double& t) const;
     
     [[nodiscard]] virtual Vector2 GetPoint(double t) const = 0;
     [[nodiscard]] virtual Vector2 GetDirection(double t) const = 0;
@@ -33,8 +33,8 @@ class LinearSegment : public EdgeSegment {
 public:
     LinearSegment(const Vector2 &p0, const Vector2 &p1);
     
-    [[nodiscard]] float Distance(const Vector2 &p, double& t) const override;
-    [[nodiscard]] float PseudoDistance(float distance, const Vector2&p, double& t) const override;
+    [[nodiscard]] double Distance(const Vector2 &p, double& t) const override;
+    [[nodiscard]] double PseudoDistance(float distance, const Vector2&p, double& t) const override;
     
     [[nodiscard]] Vector2 GetPoint(double t) const override;
     [[nodiscard]] Vector2 GetDirection(double t) const override;
@@ -45,8 +45,8 @@ class QuadraticSegment : public EdgeSegment {
 public:
     QuadraticSegment(const Vector2 &p0, const Vector2 &p1, const Vector2 &p2);
     
-    [[nodiscard]] float Distance(const Vector2 &p, double& t) const override;
-    [[nodiscard]] float PseudoDistance(float distance, const Vector2&p, double& t) const override;
+    [[nodiscard]] double Distance(const Vector2 &p, double& t) const override;
+    [[nodiscard]] double PseudoDistance(float distance, const Vector2&p, double& t) const override;
     
     [[nodiscard]] Vector2 GetPoint(double t) const override;
     [[nodiscard]] Vector2 GetDirection(double t) const override;
@@ -57,8 +57,8 @@ class CubicSegment : public EdgeSegment {
 public:
     CubicSegment(const Vector2 &p0, const Vector2 &p1, const Vector2 &p2, const Vector2 &p3);
     
-    [[nodiscard]] float Distance(const Vector2 &p, double& t) const override;
-    [[nodiscard]] float PseudoDistance(float distance, const Vector2&p, double& t) const override;
+    [[nodiscard]] double Distance(const Vector2 &p, double& t) const override;
+    [[nodiscard]] double PseudoDistance(float distance, const Vector2&p, double& t) const override;
     
     [[nodiscard]] Vector2 GetPoint(double t) const override;
     [[nodiscard]] Vector2 GetDirection(double t) const override;
